test(add): add tests for sidebar chats
