+++ tests/specs/reusable-accounts/ChatUserB/03-sidebar-with-B.spec.ts
@@ -1,0 +1,42 @@
+import ChatsLayout from "../../../screenobjects/chats/ChatsLayout";
+import InputBar from "../../../screenobjects/chats/InputBar";
+import Topbar from "../../../screenobjects/chats/Topbar";
+import FriendsScreen from "../../../screenobjects/friends/FriendsScreen";
+
+export default async function sidebarWithUserB() {
+  it("Chat User B - Wait until receiving a friend request again", async () => {
+    // Go to pending requests list, wait for receiving the friend request and accept it
+    await FriendsScreen.goToPendingFriendsList();
+    await FriendsScreen.waitUntilFriendRequestIsReceived();
+  });
+
+  it("Chat User B - Validate button badge displays the number of incoming requests", async () => {
+    await FriendsScreen.acceptIncomingRequest("ChatUserA");
+
+    // Return to Friends List
+    await FriendsScreen.goToAllFriendsList();
+  });
+
+  it("Chat User B - Accept incoming request", async () => {
+    await FriendsScreen.acceptIncomingRequest("ChatUserA");
+
+    // Return to Friends List
+    await FriendsScreen.goToAllFriendsList();
+  });
+
+  it("Send message to User B", async () => {
+    // Go to the current list of All friends and then open a Chat conversation with ChatUserA
+    await FriendsScreen.chatWithFriendButton.waitForExist();
+    await FriendsScreen.chatWithFriendButton.click();
+    await ChatsLayout.waitForIsShown(true);
+    await Topbar.waitUntilRemoteUserIsOnline();
+
+    // Send message to Chat User B
+    await InputBar.typeMessageOnInput("Helloagain...");
+    await InputBar.clickOnSendMessage();
+  });
+
+  after(async () => {
+    await browser.pause(60000);
+  });
+}
